pnp-diffusers/requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.25.0
diffusers==0.17.1
xformers==0.0.20
transformers==4.30.2
accelerate==0.20.3
gradio==4.14.0
pillow>=9.0.0
pyyaml>=6.0.0